+++ src/png/consts.rs
@@ -5,7 +5,7 @@ pub const IHDR_CHUNK_TYPE: &[u8] = &[0x49, 0x48, 0x44, 0x52];
 pub const IHDR_DATA_LENGTH: usize = 13;
 
 pub const IDAT_CHUNK_TYPE: &[u8] = &[0x49, 0x44, 0x41, 0x54];
-pub const IDAT_CHUNK_MAX_SIZE: u32 = (2 as u32).pow(31) - 1;
+pub const IDAT_CHUNK_MAX_SIZE: u32 = (2u32).pow(31) - 1;
 
 pub const IEND_CHUNK_TYPE: &[u8] = &[0x49, 0x45, 0x4e, 0x44];
